refactor: add more modules, move some functions and refactor run_processing
